requirements.txt
# 核心依赖
python>=3.8
httpx>=0.24.0
playwright>=1.32.0
undetected-chromedriver>=3.4.6
pysocks>=1.7.1
sqlite3
cryptography>=39.0.0

# 智能解析
dukpy>=0.2.3
beautifulsoup4>=4.11.0
lxml>=4.9.0

# 验证码处理
pytesseract>=0.3.10
Pillow>=9.4.0

# 网络情报收集
requests>=2.28.0
urllib3>=1.26.0

# 数据分析与处理
pandas>=1.5.0
numpy>=1.24.0

# 分布式架构
twisted>=22.4.0
redis>=4.5.0

# 工具库
python-dotenv>=1.0.0
pyyaml>=6.0